+++ src/chunk.rs
@@ -8,7 +8,7 @@ pub const SIZE: u8 = 16;
 #[derive(PartialEq)]
 pub struct Chunk {
     pub position: Point3<i32>,
-    pub blocks: [[[u16; SIZE as usize]; SIZE as usize]; SIZE as usize],
+    pub blocks: [[[u32; SIZE as usize]; SIZE as usize]; SIZE as usize],
 }
 
 impl Chunk {
@@ -18,14 +18,14 @@ impl Chunk {
             blocks: [[[0; SIZE as usize]; SIZE as usize]; SIZE as usize],
         }
     }
-    pub fn new_filled(pos: Point3<i32>, block: u16) -> Chunk {
+    pub fn new_filled(pos: Point3<i32>, block: u32) -> Chunk {
         Chunk {
             position: pos * SIZE as i32,
             blocks: [[[block; SIZE as usize]; SIZE as usize]; SIZE as usize],
         }
     }
-    pub fn new_layered(pos: Point3<i32>, surface: u16, shallow: u16, deep: u16) -> Chunk {
-        let mut blocks: [[[u16; SIZE as usize]; SIZE as usize]; SIZE as usize] =
+    pub fn new_layered(pos: Point3<i32>, surface: u32, shallow: u32, deep: u32) -> Chunk {
+        let mut blocks: [[[u32; SIZE as usize]; SIZE as usize]; SIZE as usize] =
             [[[0; SIZE as usize]; SIZE as usize]; SIZE as usize];
         for x in 0..blocks.len() {
             for y in 0..blocks[x].len() {
@@ -45,8 +45,8 @@ impl Chunk {
             blocks,
         }
     }
-    pub fn new_perlin(pos: Point3<i32>, block: u16) -> Chunk {
-        let mut blocks: [[[u16; SIZE as usize]; SIZE as usize]; SIZE as usize] =
+    pub fn new_perlin(pos: Point3<i32>, block: u32) -> Chunk {
+        let mut blocks: [[[u32; SIZE as usize]; SIZE as usize]; SIZE as usize] =
             [[[0; SIZE as usize]; SIZE as usize]; SIZE as usize];
         for x in 0..blocks.len() {
             for y in 0..blocks[x].len() {
@@ -69,23 +69,23 @@ impl Chunk {
         }
     }
 
-    pub fn get_block(&self, pos: Point3<u8>) -> u16 {
+    pub fn get_block(&self, pos: Point3<u8>) -> u32 {
         if pos.x >= SIZE || pos.y >= SIZE || pos.z >= SIZE {
             return 0;
         }
         self.blocks[pos.x as usize][pos.y as usize][pos.z as usize]
     }
-    pub fn get_block_s(&self, pos: Point3<i16>) -> u16 {
+    pub fn get_block_s(&self, pos: Point3<i16>) -> u32 {
         if pos.x >= SIZE as i16 || pos.y >= SIZE as i16 || pos.z >= SIZE as i16 || pos.x < 0 || pos.y < 0 || pos.z < 0 {
             return 0;
         }
         self.blocks[pos.x as usize][pos.y as usize][pos.z as usize]
     }
 
-    pub fn set_block(&mut self, pos: Point3<u8>, block: u16) {
+    pub fn set_block(&mut self, pos: Point3<u8>, block: u32) {
         self.blocks[pos.x as usize][pos.y as usize][pos.z as usize] = block;
     }
-    pub fn set_block_global(&mut self, pos: Point3<i32>, block: u16) {
+    pub fn set_block_global(&mut self, pos: Point3<i32>, block: u32) {
         let x = (pos.x - self.position.x) + SIZE as i32 / 2;
         let y = (pos.y - self.position.y) + SIZE as i32 / 2;
         let z = (pos.z - self.position.z) + SIZE as i32 / 2;
@@ -94,7 +94,7 @@ impl Chunk {
         };
         self.blocks[(x as i32) as usize][(y as i32) as usize][(z as i32) as usize] = block;
     }
-    pub fn get_block_global(&self, pos: Point3<i32>) -> u16 {
+    pub fn get_block_global(&self, pos: Point3<i32>) -> u32 {
         let x = (pos.x - self.position.x) + SIZE as i32 / 2;
         let y = (pos.y - self.position.y) + SIZE as i32 / 2;
         let z = (pos.z - self.position.z) + SIZE as i32 / 2;
